Rename interpolation_utils.py to interpolation.py
